+++ protected/modules/Service/controllers/OldController.php
@@ -22,16 +22,22 @@ class OldController extends GController {
      * @param type $id 客户id
      */
     public function actionEdit($id) {
-        $noteinfo = new NoteInfo();
-        $noteinfo->setAttribute("iskey", 0);
+        $noteinfo = new NoteInfo(); 
         $noteinfo->setAttribute("isvalid", 1);
         $noteinfo->setAttribute("dial_id", 0);
         $noteinfo->setAttribute("message_id", 0);
         $noteinfo->setAttribute("next_contact", '');
         $noteinfo->setAttribute("cust_id", $id);
         $model = $this->loadModel($id);
+        $noteinfo->setAttribute("iskey", $model->iskey);
         $sql = "select * from {{aftermarket_cust_info}} where cust_id=:cust_id";
         $aftermodel = AftermarketCustInfo::model()->findBySql($sql, array(':cust_id' => $id));
+        if($aftermodel->cust_type==8){
+            //公海资源
+            Yii::app()->user->setFlash('success','不能打开公海资源!');
+            $this->render("result");
+            return ;
+        }
         $sql = "select * from {{contract_info}} where cust_id=:cust_id";
         $contractmodel = ContractInfo::model()->findBySql($sql, array(':cust_id' => $id));
         $model->setAttribute("create_time", date("Y-m-d H:i:s", $model->getAttribute("create_time")));
@@ -70,18 +76,19 @@ class OldController extends GController {
      * @param integer $id the ID of the model to be updated
      */
     public function actionUpdate($id) {
-        $noteinfo = new NoteInfo();
-        $noteinfo->setAttribute("iskey", 0);
+        $noteinfo = new NoteInfo(); 
         $noteinfo->setAttribute("isvalid", 1);
         $noteinfo->setAttribute("dial_id", 0);
         $noteinfo->setAttribute("message_id", 0);
         $noteinfo->setAttribute("next_contact", '');
         $noteinfo->setAttribute("cust_id", $id);
         $model = $this->loadModel($id);
+        $noteinfo->setAttribute("iskey", $model->iskey);
         $sql = "select * from {{aftermarket_cust_info}} where cust_id=:cust_id";
         $aftermodel = AftermarketCustInfo::model()->findBySql($sql, array(':cust_id' => $id));
         $sql = "select * from {{contract_info}} where cust_id=:cust_id";
         $contractmodel = ContractInfo::model()->findBySql($sql, array(':cust_id' => $id));
+        $nextcontact='';
         if (isset($_POST['AftermarketCustInfo'])) {
             //保存   
             $newCustType = $_POST['AftermarketCustInfo']['cust_type'];
@@ -107,8 +114,9 @@ class OldController extends GController {
             //保存合同信息
             $contractmodel->save();
             //保存小记信息
-            if (Utils::isNeedSaveNoteInfo($_POST['NoteInfo'])) {
+            if (Utils::isNeedSaveNoteInfoForAfter($_POST['NoteInfo'],$newCustType)) {
                 //保存小记 
+                $nextcontact=$_POST['NoteInfo']['next_contact'];
                 $noteinfo->attributes = $_POST['NoteInfo'];
                 if ($model->iskey != $noteinfo->iskey) {
                     $model->iskey = $noteinfo->iskey;
@@ -140,14 +148,22 @@ class OldController extends GController {
             }
             if ($newCustType == 8) {
                 //客户分类转成8，生成公海资源数据
-                $blackinfo = new BlackInfo();
-                $blackinfo->setAttribute("cust_id", $id);
-                $blackinfo->setAttribute('lib_type', 3);
-                $blackinfo->setAttribute('old_cust_type', $aftermodel->cust_type);
-                $blackinfo->setAttribute('create_time', time());
-                $blackinfo->setAttribute('creator', Yii::app()->user->id);
-                $blackinfo->save();
-                $aftermodel->delete(); //删除售后库
+                if (empty($temp)) {
+                    $blackinfo = new BlackInfo();
+                    $blackinfo->setAttribute("cust_id", $id);
+                    $blackinfo->setAttribute('lib_type', 3);
+                    $blackinfo->setAttribute('old_cust_type', $aftermodel->cust_type);
+                    $blackinfo->setAttribute('create_time', time());
+                    $blackinfo->setAttribute('creator', Yii::app()->user->id);
+                    $blackinfo->save();
+                    $aftermodel->delete(); //删除售后库
+                } else {
+                    Yii::app()->db->createCommand()->update('{{black_info}}', array('lib_type' => 3,
+                            'old_cust_type' => $aftermodel->cust_type,
+                            'create_time' => time(),
+                            'creator' => Yii::app()->user->id
+                             ), "cust_id=$id");
+                }
                 $model->status = "1";
                 //售后员已分配资源数减1
                 $sql = "update {{users}} set cust_num=cust_num-1 where eno='{$aftermodel->eno}'";
@@ -170,15 +186,19 @@ class OldController extends GController {
                 $transaction->commit();
                 if ($newCustType == 8) {
                     //转入成功页面
+                    Yii::app()->user->setFlash('success',"成功放入公海!");
                     $this->render("result");
                     return;
                 } else {
                     //保存成功，没有错误，跳转到编辑页面
-                    $this->redirect($this->createUrl("old/Edit", array('id' => $id)));
+                    Yii::app()->user->setFlash('success',"保存成功!");
+                    $this->redirect($this->createUrl("edit", array('id' => $id)));
                 }
             } else {
                 $transaction->rollback();
-                $noteinfo->setAttribute("next_contact", '');
+                if(!empty($nextcontact)){
+                    $noteinfo->setAttribute("next_contact", $nextcontact);
+                }
             }
         }
         $model->setAttribute("create_time", date("Y-m-d H:i:s", $model->getAttribute("create_time")));
@@ -512,5 +532,7 @@ class OldController extends GController {
     public function getUserArr2($id) {
         return CHtml::listData(Users::model()->findAll('id=:id', array(':id' => $id)), 'id', 'name');
     }
-
+    public function genNoteRecordInfo($data){
+        return Utils::genNoteDisplayRecord($data);
+    }
 }
